requirements.txt
# Audio and speech recognition
pyaudio>=0.2.11
vosk>=0.3.44

# UI and visualization
pygame>=2.1.0
streamlit>=1.28.0
gradio>=3.50.0

# Screenshot and image processing
Pillow>=8.3.0
requests>=2.25.0

# Utilities
numpy>=1.20.0

# Local AI and environment
ollama>=0.1.0
python-dotenv>=1.0.0

# File processing
python-magic>=0.4.27

# Web UI components
streamlit-webrtc>=0.47.0
streamlit-chat>=0.1.1 